requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
alluka==0.1.4
attrs==23.2.0
banal==1.0.6
beautifulsoup4>=4.12.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorlog==6.8.2
dataset>=1.6.2
easygoogletranslate>=0.0.2
frozenlist==1.4.1
hikari>=2.0.0.dev120
hikari-lightbulb>=2.3.3
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
python-dotenv>=1.0.0
requests>=2.31.0
SQLAlchemy==1.4.51
typing_extensions==4.9.0
urllib3==1.26.18
yarl==1.9.4